requirements.txt
numpy==2.2.3
opencv_python_headless==4.11.0.86
pandas==2.3.3
openpyxl==3.1.5
Pillow==12.1.0
pytesseract==0.3.13
rapidfuzz==3.14.3
torch==2.8.0
torchvision==0.23.0
transformers==4.57.3
ultralytics==8.3.236